+++ src/hooman.d.ts
@@ -1,0 +1,4 @@
+declare module "hooman" {
+  const got: import("got").Got;
+  export default got;
+}