+++ tests/author.test.ts
@@ -1,13 +1,17 @@
-import { Login, Gallery } from "../src/index";
+import { Login, Author, Gallery, user, User } from "../src/index";
 import * as config from "./jest.config";
-const { userId } = config.options.authorOptions;
 const { cookieA, cookieB } = config.options.loginOptions;
 
 test("Test author, gallery and scraps", async () => {
   Login(cookieA, cookieB);
-  //   // author
-  //   const results = await Author(userId)
-  //   expect(!!results.id).not.toBeFalsy()
+  const userId = "recallfuture";
+  // author
+  const author = await Author(userId)
+  expect(author.id).toBe(userId)
+
+  const user = await User()
+  expect(user.id).toBe(userId)
+
   // gallery
   const gallery = await Gallery(userId, 1);
   expect(gallery instanceof Array).toBeTruthy();